// block-performance-1.0.0/src/library/web-client/post-json/type.ts
export interface Input {
    url: string;
    url_params?: { [key: string]: string };
    headers?: { [key: string]: string };
    proxy_url: string;
    data: any;
}

export interface OutputOk {
    json_value: any;
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}